Ensure that '/var/run/resource-agents' is created
When the cluster resources are started, the Resource Agents expect '/var/run/resource-agents/'
to exists. However, the RA's nor any other piece of software creates the directory. This has
been discussed upstream and it was recommended [1] to verify/create such directory on the init
scripts (corosync/heartbeat/pacemaker mcp)

[1]: http://developerbugs.linux-foundation.org/show_bug.cgi?id=2378#c17

LP: #787062
Closes: #627685