Comment in quorum provider.
If we don't enable corosync_votequorum pacemaker will claim about the absence
about a quorum provider.

We need to add any value for expected_votes, otherwise corosync will fail to
start. However we can't determine the size of users cluster, so the example
config uses 2. This is the smallest reasonable size for cluster.

Signed-off-by: Adrian Vondendriesch <adrian.vondendriesch@credativ.de>