Freshen package descriptions
Corosync 2 is not an OpenAIS implementation like Corosync 1 was.
Replace the long descriptions with the current welcome text from the homepage.
To keep them (and the short descriptions as well) in sync, use control file
substitutions.  The deb-substvars syntax is rather awkward to work with
directly, so generate it from a plain text description.