Make sure environment variables are propagated by the init script
We use the default file to set Corosync-specific environment variables,
which are not used by the init script, but have to be exported to the
started daemon.  The systemd service also uses this file via an
EnvironmentFile directive, which does not tolerate the "export" keywords,
so a shell workaround was chosen instead.  While this needlessly exports
the OPTIONS variable, this is not expected to cause problems.