Really add arm64 support
The autogenerated GitHub archive, which we used as orig tarball for 2.3.4,
did not contain the m4 directory.  This broke the default dh_autoreconf
action (that is, autoreconf -f -i), and we chose the easy way out: invoking
the supplied autogen.sh.  Unfortunately, that invokes autoreconf without
the -f option, thus stopped updating config.sub and config.guess when used
on proper distribution tarballs (like that of 2.3.5).

Closes: #753289